Add fuzz test
